SupplyChain-Django-backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.2
Deprecated==1.2.13
Django==4.0
djangorestframework==3.14.0
fonttools==4.28.5
frozenlist==1.2.0
gcloud==0.18.3
googleapis-common-protos==1.54.0
hexbytes==0.2.2
httplib2==0.20.2
idna==3.3
ipfshttpclient==0.8.0a2
joblib==1.1.0
jsonschema==3.2.0
jwcrypto==1.0
kiwisolver==1.3.2
lru-dict==1.1.7
matplotlib==3.5.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
numpy==1.21.5
oauth2client==4.1.3
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
parsimonious==0.8.1
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pyparsing==3.0.6
PyQRCode==1.2.1
Pyrebase4==4.5.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-jwt==3.3.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
rlp==2.0.1
rsa==4.8
scikit-learn==0.22.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
varint==1.0.2
web3==5.25.0
websockets==9.1
wrapt==1.13.3
xgboost==1.6.0
yarl==1.7.2
gunicorn==20.1.0